refactor tests

+++ src/modules/address/services/delete-address-service.ts
@@ -1,3 +1,5 @@
+import 'reflect-metadata';
+import { AppError } from './../../../shared/errors/AppError';
 import { IAddressRepository } from './../repositories/IAddressRepository';
 import { inject, injectable } from 'tsyringe';
 
@@ -8,6 +10,11 @@ export class DeleteAddressService {
   ) {}
 
   public async execute(address_id: string): Promise<void> {
+    const address = await this.addressRepository.findById(address_id);
+
+    if (!address) {
+      throw new AppError('Address is not Found', 401);
+    }
     await this.addressRepository.delete(address_id);
   }
 }
